# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Executive Assessment API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AssessmentInput {
  firstName: string;
  lastName: string;
  email: string;
  company: string;
  jobTitle: string;
  industry?: string;
}

export interface AssessmentUpdate {
  firstName?: string;
  lastName?: string;
  email?: string;
  company?: string;
  jobTitle?: string;
  industry?: string;
}

export type AssessmentStatus = typeof AssessmentStatus[keyof typeof AssessmentStatus];


export const AssessmentStatus = {
  draft: 'draft',
  submitted: 'submitted',
  paid: 'paid',
  approved: 'approved',
  completed: 'completed',
} as const;

export interface Assessment {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  company: string;
  jobTitle: string;
  /** @nullable */
  industry?: string | null;
  status: AssessmentStatus;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export type SectionResponseInputData = { [key: string]: unknown };

export interface SectionResponseInput {
  data: SectionResponseInputData;
}

export type SectionResponseData = { [key: string]: unknown };

export interface SectionResponse {
  id: number;
  assessmentId: number;
  sectionNum: number;
  data: SectionResponseData;
}

export interface AssessmentDetail {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  company: string;
  jobTitle: string;
  /** @nullable */
  industry?: string | null;
  status: string;
  sections: SectionResponse[];
  /** @nullable */
  results?: string | null;
  /** @nullable */
  roadmap?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface AssessmentSummary {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  company: string;
  jobTitle: string;
  /** @nullable */
  industry?: string | null;
  status: string;
  sectionsCompleted?: number;
  createdAt: string;
}

export interface CheckoutSession {
  checkoutUrl: string;
  sessionId: string;
}

export interface ResultsInput {
  /** Markdown or rich text executive summary/findings */
  results: string;
  /** 30-60-90 day roadmap text */
  roadmap: string;
}

export interface AdminStats {
  total: number;
  draft: number;
  submitted: number;
  paid: number;
  approved: number;
  completed: number;
}

export type ListAdminAssessmentsParams = {
status?: string;
};