# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Executive Assessment API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  total: number;
  draft: number;
  submitted: number;
  paid: number;
  approved: number;
  completed: number;
}