# lib/api-zod/src/generated/types/assessment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Executive Assessment API
 * OpenAPI spec version: 0.1.0
 */
import type { AssessmentStatus } from './assessmentStatus';

export interface Assessment {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  company: string;
  jobTitle: string;
  /** @nullable */
  industry?: string | null;
  status: AssessmentStatus;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}