# lib/api-zod/src/generated/types/assessmentDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Executive Assessment API
 * OpenAPI spec version: 0.1.0
 */
import type { SectionResponse } from './sectionResponse';

export interface AssessmentDetail {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  company: string;
  jobTitle: string;
  /** @nullable */
  industry?: string | null;
  status: string;
  sections: SectionResponse[];
  /** @nullable */
  results?: string | null;
  /** @nullable */
  roadmap?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}