# lib/api-zod/src/generated/types/assessmentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Executive Assessment API
 * OpenAPI spec version: 0.1.0
 */

export interface AssessmentInput {
  firstName: string;
  lastName: string;
  email: string;
  company: string;
  jobTitle: string;
  industry?: string;
}