# lib/api-zod/src/generated/types/assessmentSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Executive Assessment API
 * OpenAPI spec version: 0.1.0
 */

export interface AssessmentSummary {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  company: string;
  jobTitle: string;
  /** @nullable */
  industry?: string | null;
  status: string;
  sectionsCompleted?: number;
  createdAt: string;
}