# lib/api-zod/src/generated/types/checkoutSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Executive Assessment API
 * OpenAPI spec version: 0.1.0
 */

export interface CheckoutSession {
  checkoutUrl: string;
  sessionId: string;
}