# lib/api-zod/src/generated/types/resultsInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Executive Assessment API
 * OpenAPI spec version: 0.1.0
 */

export interface ResultsInput {
  /** Markdown or rich text executive summary/findings */
  results: string;
  /** 30-60-90 day roadmap text */
  roadmap: string;
}