# lib/api-zod/src/generated/types/sectionResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Executive Assessment API
 * OpenAPI spec version: 0.1.0
 */
import type { SectionResponseData } from './sectionResponseData';

export interface SectionResponse {
  id: number;
  assessmentId: number;
  sectionNum: number;
  data: SectionResponseData;
}