# lib/api-zod/src/generated/types/sectionResponseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Executive Assessment API
 * OpenAPI spec version: 0.1.0
 */
import type { SectionResponseInputData } from './sectionResponseInputData';

export interface SectionResponseInput {
  data: SectionResponseInputData;
}